requirements.txt
aiogram==3.20.0.post0
openai==1.88.0
SQLAlchemy==2.0.41
aiosqlite==0.21.0
httpx==0.28.1
loguru==0.7.3
python-dotenv==1.1.0 